Changed repository class to interface
