+++ app/src/main/java/ru/myitschool/todo/ui/MainActivity.kt
@@ -2,18 +2,13 @@ package ru.myitschool.todo.ui
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.app.AppCompatDelegate
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
-import androidx.work.OneTimeWorkRequestBuilder
-import androidx.work.WorkManager
 import com.google.android.material.snackbar.Snackbar
-import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
 import ru.myitschool.todo.App
 import ru.myitschool.todo.R
-import ru.myitschool.todo.data.repository.SharedPreferencesRepository
 import ru.myitschool.todo.databinding.ActivityMainBinding
 import ru.myitschool.todo.utils.NetworkStateMonitor
 import javax.inject.Inject
