+++ app/src/main/java/ru/myitschool/todo/ui/settings_fragment/view/SettingsFragment.kt
@@ -25,6 +25,7 @@ import kotlinx.coroutines.launch
 import ru.myitschool.todo.App
 import ru.myitschool.todo.R
 import ru.myitschool.todo.data.repository.SharedPreferencesRepository
+import ru.myitschool.todo.data.repository.impl.SharedPreferencesRepositoryImpl
 import ru.myitschool.todo.databinding.FragmentSettingsBinding
 import ru.myitschool.todo.di.components.AppComponent
 import ru.myitschool.todo.ui.ViewModelFactory
