+++ app/src/main/java/ru/myitschool/todo/utils/NetworkListener.kt
@@ -5,13 +5,15 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
 import ru.myitschool.todo.data.repository.TodoItemsRepository
+import ru.myitschool.todo.data.repository.impl.TodoItemsRepositoryImpl
 import ru.myitschool.todo.di.scopes.AppScope
 import javax.inject.Inject
 
 @AppScope
 class NetworkListener @Inject constructor(
     private val networkStateMonitor: NetworkStateMonitor,
-    private val repository: TodoItemsRepository) {
+    private val repository: TodoItemsRepository
+) {
     private val scope = CoroutineScope(Dispatchers.Main)
     private var currentJob: Job? = null
     fun startListener(){
